Added automatic naming for output files.
LIME can now automatically name output files using the number of grid points (par->pIntensity) and sink points
(par->sinkPoints) and an appropriate file extension. This happens when a configInfo filename parameter is left
blank (i.e. ""). The user can also set a custom prefix to be added to every output filename by setting par
>fileprefix and including "[prefix]" in a configInfo filename parameter. If "[prefix]" is included then par-
>filePrefix must be set. Note if a configInfo parameter is not set then no automatic naming is performed and the
file is not outputted as normal. The example model file has been updated with this information and an example to
help explain the concept.

This automatic naming has also been extended to the images using the same logic. Included in the automatic naming is the number of grid points (par->pIntensity) and sink points (par->sinkPoints) as well as the image inclination, position angle and azimuthal information. The frequency (img[].freq) is also added (in GHz) if the image is in continuum, otherwise for a line image the molecule (extracted from moldatfile[0]) and transition (img[].trans) is added. The unit type is converted to text and appended and then finally a .fits extension is added.

The above has been achieved by adding some new functions into aux.c: *makeFilename, *completeFilename, *completeImageFilename, *addExtension, *removeExtension, *replacePrefix. Note also that the strlen(inStr)<=0 condition has been removed from copyInparStr() so that NULL and empty strings can now be differentiated.